+++ src/io/output/FileWriter.h
@@ -22,6 +22,6 @@ class FileWriter {
      * @param iteration The current iteration number
      * @param particle_container ParticleContainer which provides the data to be written
      */
-    virtual void writeFile(const SimulationParams& params, int iteration,
+    virtual void writeFile(const SimulationParams& params, size_t iteration,
                            const std::unique_ptr<ParticleContainer>& particle_container) const = 0;
 };
